refactor: separate workspace setup notice
